+++ mesa/experimental/__init__.py
@@ -1,7 +1,4 @@
 """Experimental init."""
-
-from mesa.experimental import cell_space
-
 from .solara_viz import JupyterViz, Slider, SolaraViz, make_text
 
-__all__ = ["cell_space", "JupyterViz", "SolaraViz", "make_text", "Slider"]
+__all__ = ["JupyterViz", "SolaraViz", "make_text", "Slider"]
