--- mesa/experimental/cell_space/__init__.py
@@ -1,32 +1,0 @@
-"""Cell spaces.
-
-Cell spaces offer an alternative API for discrete spaces. It is experimental and under development. The API is more
-expressive that the default grids available in `mesa.space`.
-
-"""
-
-from mesa.experimental.cell_space.cell import Cell
-from mesa.experimental.cell_space.cell_agent import CellAgent
-from mesa.experimental.cell_space.cell_collection import CellCollection
-from mesa.experimental.cell_space.discrete_space import DiscreteSpace
-from mesa.experimental.cell_space.grid import (
-    Grid,
-    HexGrid,
-    OrthogonalMooreGrid,
-    OrthogonalVonNeumannGrid,
-)
-from mesa.experimental.cell_space.network import Network
-from mesa.experimental.cell_space.voronoi import VoronoiGrid
-
-__all__ = [
-    "CellCollection",
-    "Cell",
-    "CellAgent",
-    "DiscreteSpace",
-    "Grid",
-    "HexGrid",
-    "OrthogonalMooreGrid",
-    "OrthogonalVonNeumannGrid",
-    "Network",
-    "VoronoiGrid",
-]
