+++ mesa/spaces/__init__.py
@@ -1,0 +1,20 @@
+from mesa.spaces.network import Network
+from mesa.spaces.cell import Cell
+from mesa.spaces.cell_agent import CellAgent
+from mesa.spaces.cell_collection import CellCollection
+from mesa.spaces.discrete_space import DiscreteSpace
+from mesa.spaces.grid import Grid, HexGrid, OrthogonalMooreGrid, OrthogonalVonNeumannGrid
+from mesa.spaces.voronoi import VoronoiGrid
+
+__all__ = [
+    "CellCollection",
+    "Cell",
+    "CellAgent",
+    "DiscreteSpace",
+    "Grid",
+    "HexGrid",
+    "OrthogonalMooreGrid",
+    "OrthogonalVonNeumannGrid",
+    "Network",
+    "VoronoiGrid",
+]
