+++ mesa/spaces/cell_collection.py
@@ -9,8 +9,8 @@ from random import Random
 from typing import TYPE_CHECKING, Generic, TypeVar
 
 if TYPE_CHECKING:
-    from mesa.experimental.cell_space.cell import Cell
-    from mesa.experimental.cell_space.cell_agent import CellAgent
+    from mesa.spaces.cell import Cell
+    from mesa.spaces.cell_agent import CellAgent
 
 T = TypeVar("T", bound="Cell")
